requirements.txt
requests==2.22.0
Flask==1.1.1
telebot==0.0.3
pyTelegramBotAPI==3.6.6
psycopg2-binary==2.8.4
beautifulsoup4==4.8.1
python-dotenv==0.12.0
